// LiveTalk-main/types.ts
export interface VIPPackage {
  id?: string;
  level: number;
  name: string;
  cost: number;
  color: string;
  frameUrl: string;
  nameStyle: string;
  privileges?: string[]; // قائمة المميزات
  durationDays?: number; // مدة الاشتراك بالأيام
}